make some edit in invoices

+++ app/Http/Controllers/InvoicesController.php
@@ -180,14 +180,24 @@ class InvoicesController extends Controller
         $invoices = invoices::where('id', $id)->first();
         $invoice_attachments = invoice_attachments::where('invoice_attachment_id', $id)->first();
 
-        if(!empty($invoice_attachments->invoice_number)) {
-            // Storage::disk('public_uploads')->delete($invoice_attachments->invoice_number . "/" . $invoice_attachments->file_name);  /////  this code if i want delete file inside folder
-            Storage::disk('public_uploads')->deleteDirectory($invoice_attachments->invoice_number); ///// this code if i want delete all folder
-        }
+        $id_page = $request->id_page;
+
+        // to check if modal is archive_invoices or delete_invoices
+        if($id_page == 1) {
+            if(!empty($invoice_attachments->invoice_number)) {
+                // Storage::disk('public_uploads')->delete($invoice_attachments->invoice_number . "/" . $invoice_attachments->file_name);  /////  this code if i want delete file inside folder
+                Storage::disk('public_uploads')->deleteDirectory($invoice_attachments->invoice_number); ///// this code if i want delete all folder
+            }
+            $invoices->forceDelete();
+            session()->flash('delete_invoice');
+            return redirect('invoices');
+        } else {
         // i used delete() function because i want delete this invoice from table but i want make copy in database (soft delete)
         $invoices->delete();
-        session()->flash('delete_invoice');
-        return redirect('invoices');
+        session()->flash('archive_invoice');
+        return redirect('invoices_archive');
+        }
+        
     }
 
     // this function to make product from select in add invoice
